requirements.txt
autopep8>=1.5.6
binarytree>=6.3.0
black>=21.4b1
Flask>=1.1.2
locust>=1.5.3
pre-commit>=2.12.1
pylint>=2.7.2
python-dateutil>=2.8.1
PyYAML>=5.4.1
regex>=2021.4.4
repeat>=0.1.3
requests>=2.25.1
requests-oauthlib>=1.3.0
rsa>=4.7.2
